Move mock helpers to the utils folder
