+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/utils/MockSecurityAnalyticsService.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.helpers;
+package com.wazuh.contentmanager.utils;
 
 import com.fasterxml.jackson.databind.JsonNode;
